+++ src/cpp/frontend/esphome/espectre/espectre.h
@@ -70,6 +70,7 @@ class ESpectreComponent : public Component, public IRuntimeListener {
   
   // Setters for ESPHome sensors (delegated to SensorPublisher)
   void set_movement_sensor(sensor::Sensor *sensor) { this->sensor_publisher_.set_movement_sensor(sensor); }
+  void set_intensity_sensor(sensor::Sensor *sensor) { this->sensor_publisher_.set_intensity_sensor(sensor); }
   void set_motion_binary_sensor(binary_sensor::BinarySensor *sensor) { this->sensor_publisher_.set_motion_binary_sensor(sensor); }
   
   // Setter for threshold number control